chore: dedicated hook to get descriptors from the wallet config
